recipes-qt: qtbase: make gles2 support optional
gles2 support requires to add opengl to DISTRO_FEATURES which is quite
drastic and not nice to other layers which might not need it.

gles2 support in qtbase is only needed for nymea-app. Since nymea-app
explicits the dependency on opengl in DISTRO_FEATURES, we can make gles2
support in qtbase depend on the presence of opengl in DISTRO_FEATURES.

Signed-off-by: Quentin Schulz <quentin.schulz@streamunlimited.com>